Clean up and move utils to main.
