--- utils.py
@@ -1,28 +1,0 @@
-'''
-Utility functions for 7 days game
-'''
-
-import sys
-import os
-import random
-import time
-
-# add color output
-# https://github.com/tartley/colorama
-from colorama import init
-init()
-
-from colorama import Fore, Back, Style
-
-def clear_screen():
-    '''
-    Clear the terminal screen.
-    '''
-    os.system('clear')
-
-
-def game_over_because(reason):
-    '''
-    Display game over screen.
-    '''
-    print(Fore.RED + "\nGAME OVER\n===============\n" + reason  + "\nBetter luck next time!")
